04-april-2023 work

+++ StudentManagement/src/app/student/student.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild, ElementRef, OnChanges, SimpleChanges } from '@angular/core';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { APIService } from '../services/APIservice.service';
 import { StudentDetailesService } from '../services/studentDetailes';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -16,8 +16,11 @@ export class StudentComponent implements OnInit {
   studentsData:StudentDetailesService[]=[];
   studentsDataLength!:number;
   showStudentDetails!:boolean;
+  
   @ViewChild('checkStdId') checkStdId!:ElementRef;
-  constructor(private APIService:APIService, private MessageService:MessageService, private route:Router , private activatedRoute:ActivatedRoute, private VariableService:VariableService ){
+  constructor(private APIService:APIService, private MessageService:MessageService, private route:Router ,
+               private activatedRoute:ActivatedRoute, private VariableService:VariableService,
+               private ConfirmationService:ConfirmationService ){
 
   }
   
@@ -30,9 +33,16 @@ export class StudentComponent implements OnInit {
     this.checkStdId.nativeElement.value='';
   }
   getAllStudentDetailes(){
+    this.studentsData=[]
     this.APIService.findAllStudents().subscribe((results)=>{
-      this.studentsData=Object.values(results);
-      console.log(results);
+      for(let result of Object.values(results)){
+        console.log(result);
+        
+        let data=this.VariableService.getFormatedStudentData(result)
+        this.studentsData.push(data)
+        // console.log(data);
+        
+      }
     })
   }
 
@@ -50,12 +60,34 @@ export class StudentComponent implements OnInit {
     }
   }
 
-  deleteByStdId(stdId:string){
-    this.APIService.deleteByStudentId(stdId).subscribe((results)=>{ }, (error)=>{
-      console.log(error);
-      this.getAllStudentDetailes();
-    })
-  }
+  deleteByStdId(stdId:number){
+    this.ConfirmationService.confirm({
+        message: 'Are you want to delete the record?',
+        header: 'Confirmation',
+        // icon: 'pi pi-exclamation-triangle',
+        accept: () => {
+          // console.log("acc");
+          
+          this.MessageService.add({severity:'success', summary:'Confirmed', detail:'Record is deleted successfully'});
+          this.APIService.deleteByStudentId(stdId).subscribe((results)=>{ }, (error)=>{
+            console.log(error);
+            this.getAllStudentDetailes();
+          })
+        },
+        reject: () => {
+          // console.log("no");
+          
+          this.MessageService.add({severity:'error', summary:'Rejected', detail:'Record is not Deleted'});
+        }
+    });
+}
+
+  // deleteByStdId(stdId:string){
+  //   this.APIService.deleteByStudentId(stdId).subscribe((results)=>{ }, (error)=>{
+  //     console.log(error);
+  //     this.getAllStudentDetailes();
+  //   })
+  // }
   studentDetailesEdit(studentId:string){
     this.route.navigate(['StudentDetails/:'+studentId+''],{relativeTo:this.activatedRoute})
     this.showStudentDetails=false;
